1.add life cycle check for twonky share folder 2.improve notification of local file copy/move 3.remove unecessary version check 4.adjust event notify loader flow

+++ app/src/main/java/com/transcend/nas/management/SmbFileDeleteLoader.java
@@ -44,6 +44,11 @@ public class SmbFileDeleteLoader extends SmbAbstractLoader {
 
     private boolean delete() throws MalformedURLException, SmbException {
         for (String path : mPaths) {
+            if(isLoadInBackgroundCanceled()) {
+                Log.d(TAG, "Delete cancel");
+                break;
+            }
+
             SmbFile target = new SmbFile(getSmbUrl(path));
             target.delete();
         }
